Extract catalog search surfaces
